moved settings
