+++ src/components/TabBar/SettingsDrawerStyles.ts
@@ -2,10 +2,15 @@ import { createStyles, Theme } from '@material-ui/core/styles';
 
 export const styles = (theme: Theme) =>
   createStyles({
+    container: {
+      paddingTop: 40,
+      margin: 24,
+      height: '100vh'
+    },
     title: {
       color: 'green',
       padding: 10,
-      margin: 30
+
     },
     list: {
       width: 300
