+++ src/pages/Settings/Settings.tsx
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import SettingsDrawer from '../../components/TabBar/SettingsDrawer';
+import { Link } from 'react-router-dom';
+
+
+function Settings({ match }) {
+    console.log("HEYYYY")
+  return (
+    <div>hi
+        <SettingsDrawer />
+        <div>HJDKSFGHJKL</div>
+    </div>
+  );
+}
+
+export default Settings;
